Force lazy initialization of user sessions to use user sessions in some action without destroying the cache ability by browsers for the most other requests.
Sessions are only initialized when the user object is really needed by an action. This is done by »proxy« classes for
storage and user and a extended context. The real storage/user instances will be created at the first accessing:

CONTEXT->getStrorage();
CONTEXT->getUser();

This patch is useless if you use the setting 'core.security'. In this case the user/session is allways instantiate
by the agavi routing.