Select dialog for db related make targets.
make targets as "db" and "reindex" now ask for the DB via a
list of selctable options. You no longer need to type the name on a DB and look it up
in databases.xml

Also introduces Pulq\Agavi\Database\PulqDatabaseManager
to allow custom methods on the DB manger level.